requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
bleach==3.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-log==0.3.2
colorama==0.4.4
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fett==0.3.2
idna==2.10
iniconfig==1.1.1
keyring==21.8.0
networkx==2.5
packaging==20.8
pathtools==0.1.2
pkginfo==1.6.1
pluggy==0.13.1
py==1.10.0
Pygments==2.7.3
pymongo==3.11.2
pyparsing==2.4.7
# Editable install with no version control (pysds011==0.0.2)
-e c:\userdata\draft\dust\pysds011\src
pyserial==3.5
pytest==6.2.1
pytest-mock==3.5.0
python-jsonrpc-server==0.3.4
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==28.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rstcheck==3.3.1
six==1.15.0
snooty-lextudio==1.8.6.dev0
toml==0.10.2
tqdm==4.55.1
twine==3.3.0
typing-extensions==3.7.4.3
urllib3==1.25.11
watchdog==0.10.4
webencodings==0.5.1
